backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
httpx==0.28.1
openai==1.57.2
pydantic==2.10.3
python-dotenv==1.0.1
asyncio
python-dotenv